style: implement code review recommendations
