+++ src/components/dashboard/Sidebar.tsx
@@ -39,7 +39,7 @@ export default function Sidebar() {
     <div className='col-span-3 border border-gray-200 rounded-md shadow-[0px_2px_8px_ 0px_rgba(39, 41, 55, 0.08)] p-6 pt-9 min-h-[733px]'>
       <div className='relative rounded-full w-[120px] h-[120px] mx-auto mb-9'>
         <Image src={DefaultUserImg} width={120} height={120} alt='Journey Bites' className='rounded-full'/>
-        <Button size='icon' className='group absolute -right-3 bottom-1'>
+        <Button variant='icon' className='group absolute -right-3 bottom-1'>
           <Camera size ={20} className='group-hover:stroke-primary-100 stroke-primary'/>
         </Button>
       </div>
