+++ src/types/index.ts
@@ -1,12 +1,6 @@
-import type { ReactNode } from 'react';
-
 export enum InputType {
   TEXT = 'text',
   PASSWORD = 'password',
-}
-
-export interface BaseLayoutProps {
-  children: ReactNode
 }
 
 export type Tab = {
